+++ packages/york-react-native/src/components/RadioItem/README.md
@@ -1,0 +1,15 @@
+```js
+const ExampleComponent = () => {
+  const [value, setValue] = React.useState(null)
+
+  return (
+    <RadioItem
+      name="expamleRadioGroup"
+      isSelected={value}
+      onPress={() => setValue(true)}
+    />
+  )
+}
+
+;<ExampleComponent />
+```
